requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg[binary]
line-bot-sdk
pillow
